Feature Complete Game
This is a stable build of the game that has all of the core gameplay mechanics.
The enemy will now roam, chase or flee from the player depending on whether the player was seen or whether the player can kill the enemy.
A bomb defusing minigame which rewards the player with being able to kill the enemies for a few moves.
The enemy will now move on their own after a short amount of time to encourage the player to make faster decisions and to apply pressure when defusing the bombs.
The player now has multiple lives.

The game is still lacking sound and UI before being fully complete.